+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.1
+
+* Fix `MultiChannel`'s constructor to take a `StreamChannel`. This is
+  technically a breaking change, but since 1.0.0 was only released an hour ago,
+  we're treating it as a bug fix.
+
 ## 1.0.0
 
 * Initial version